+++ test/config_api_tests.jl
@@ -3,7 +3,7 @@ using Test
 using TrackedFloats
 
 @testset "config_* doesn't override everything" begin
-  global_config = ft__get_global_ft_config_for_test()
+  global_config = tf__get_global_tf_config_for_test()
   mirror = FtConfig(LoggerConfig(),
                     InjectorConfig(),
                     SessionConfig())